+++ src/GafferModule/MetadataAlgoBinding.cpp
@@ -65,6 +65,16 @@ void setChildNodesAreReadOnlyWrapper( Node &node, bool readOnly, bool persistent
 	setChildNodesAreReadOnly( &node, readOnly, persistent );
 }
 
+bool readOnlyWrapper( GraphComponent &g )
+{
+	return readOnly( &g );
+}
+
+GraphComponentPtr readOnlyReasonWrapper( GraphComponent &g )
+{
+	return const_cast<GraphComponent *>( readOnlyReason( &g ) );
+}
+
 void setBookmarkedWrapper( Node &node, bool bookmarked, bool persistent )
 {
 	IECorePython::ScopedGILRelease gilRelease;
@@ -120,7 +130,8 @@ void GafferModule::bindMetadataAlgo()
 	def( "getReadOnly", &getReadOnly );
 	def( "setChildNodesAreReadOnly", &setChildNodesAreReadOnlyWrapper, ( arg( "node" ), arg( "readOnly"), arg( "persistent" ) = true ) );
 	def( "getChildNodesAreReadOnly", &getChildNodesAreReadOnly );
-	def( "readOnly", &readOnly );
+	def( "readOnly", &readOnlyWrapper );
+	def( "readOnlyReason", &readOnlyReasonWrapper );
 	def(
 		"readOnlyAffectedByChange",
 		(bool (*)( const GraphComponent *, IECore::TypeId, const IECore::StringAlgo::MatchPattern &, const IECore::InternedString &, const Gaffer::Plug * ))&readOnlyAffectedByChange,